Revert "refactor(router-core): move redirect to core (#3662)"
This reverts commit 3a2d51e11d2e7a5b46f28ffc34290811bdc8a85c.
