+++ packages/router-core/src/index.ts
@@ -357,5 +357,3 @@ export type { UseLoaderDepsResult, ResolveUseLoaderDeps } from './useLoaderDeps'
 export type { UseLoaderDataResult, ResolveUseLoaderData } from './useLoaderData'
 
 export type { Redirect, ResolvedRedirect, AnyRedirect } from './redirect'
-
-export { redirect, isRedirect, isResolvedRedirect } from './redirect'
